Send service start time for serverless processes
